+++ src/main/java/chess/service/GameService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 
 import chess.database.GameStateGenerator;
-import chess.database.vanillajdbc.dao.BoardDao;
-import chess.database.vanillajdbc.dao.GameDao;
+import chess.database.dao.BoardDao;
+import chess.database.dao.GameDao;
 import chess.database.dto.BoardDto;
 import chess.database.dto.GameStateDto;
 import chess.database.dto.PointDto;
