+++ src/test/java/chess/database/dao/BoardDaoTest.java
@@ -13,10 +13,9 @@ import chess.database.dto.BoardDto;
 import chess.database.dto.GameStateDto;
 import chess.database.dto.PointDto;
 import chess.database.dto.RouteDto;
-import chess.database.vanillajdbc.dao.BoardDao;
-import chess.database.vanillajdbc.dao.JdbcBoardDao;
-import chess.database.vanillajdbc.dao.JdbcConnector;
-import chess.database.vanillajdbc.dao.JdbcGameDao;
+import chess.database.dao.vanilla.JdbcBoardDao;
+import chess.database.dao.vanilla.JdbcConnector;
+import chess.database.dao.vanilla.JdbcGameDao;
 import chess.domain.board.Board;
 import chess.domain.board.InitialBoardGenerator;
 import chess.domain.board.Point;
